split peer and related peer controllers
